# go/api_accounts.go
/*
 * AnyPay
 *
 * This the AnyPay service targeted towards, parents with children doing payments and russian oligarks
 *
 * API version: 1.0.0
 * Contact: per.von.rosen@swedbank.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// AddAccount - Add a new account
func AddAccount(c *gin.Context) {
	c.JSON(http.StatusOK, "Account added\n")
}

// GetAccount - Get a specific account with details
func GetAccount(c *gin.Context) {
	c.JSON(http.StatusOK, "There you got your specific Account\n")
}

// GetAccounts - Get all accounts IDs that the login has privileges to get
func GetAccounts(c *gin.Context) {
	c.JSON(http.StatusOK, "Yep. A list of accounts was delivered. Can you see it?? :-)\n")
}

// UpdateAccount - Update an existing account
func UpdateAccount(c *gin.Context) {
	c.JSON(http.StatusOK, "Account updated. Now it belongs to me.\n")
}