# go/api_fxorders.go
/*
 * AnyPay
 *
 * This the AnyPay service targeted towards, parents with children doing payments and russian oligarks
 *
 * API version: 1.0.0
 * Contact: per.von.rosen@swedbank.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// AddFxorder - Add a new Foreig Exchange order
func AddFxorder(c *gin.Context) {
	c.JSON(http.StatusOK, "FX Order place on market place\n")
}

// DeleteFxorder - Deletes a payment-instruction not settled yet
func DeleteFxorder(c *gin.Context) {
	c.JSON(http.StatusOK, "FX Order deleted. You were so close...\n")
}

// GetFxorder - Get all non-matched FX orders that the login has privileges to get
func GetFxorder(c *gin.Context) {
	c.JSON(http.StatusOK, "Get FX Order list. Offer:100MM EUR/SEK @ 10.53 , BID:100MM EUR/SEK @ 10.51 . \n")
}

// GetFxorder2 - Get a specific foreign exchange order (FX order) with details
func GetFxorder2(c *gin.Context) {
	c.JSON(http.StatusOK, "Get specific FX Order status. You are best offer: 100MM EUR/SEK @ 10.53 . \n")
}

// UpdateFxorder - Update a existing  FX order not matched yet, on the market place
func UpdateFxorder(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}