# go/api_users.go
/*
 * AnyPay
 *
 * This the AnyPay service targeted towards, parents with children doing payments and russian oligarks
 *
 * API version: 1.0.0
 * Contact: per.von.rosen@swedbank.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// AddUser - Add a new user
func AddUser(c *gin.Context) {
	c.JSON(http.StatusOK, "User added\n")
}

// GetUser - Get a specific user with details
func GetUser(c *gin.Context) {
	c.JSON(http.StatusOK, "Get specific User\n")
}

// GetUses - Get all user IDs that the login has privileges to get
func GetUses(c *gin.Context) {
	c.JSON(http.StatusOK, "Get list of Users\n")
}

// UpdateUser - Update an existing user
func UpdateUser(c *gin.Context) {
	c.JSON(http.StatusOK, "User updated\n")
}