# go/model_payment.go
/*
 * AnyPay
 *
 * This the AnyPay service targeted towards, parents with children doing payments and russian oligarks
 *
 * API version: 1.0.0
 * Contact: per.von.rosen@swedbank.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

/*Payment - a struct holding the Payment data necessary to IDentify and create a Payment*/
type Payment struct {
	PaymentID string `json:"paymentID,omitempty"`

	Name string `json:"name"`

	Purpose string `json:"purpose,omitempty"`

	FromAccountID int32 `json:"fromAccountID"`

	ToAccountID int32 `json:"toAccountID"`

	Value float32 `json:"value"`

	Currency string `json:"currency"`
}